requirements.txt
tensorflow
pandas 
dvc
numpy
notebook
matplotlib
seaborn
scipy
Flask
pyYAML
tqdm
joblib
python-box
ensure
ensure
-e .